🔨 Modify: handleAddToCart 함수를 ProductDetail에 직접 지정
전역 지정을 위해 하위 컴포넌트인 handleAddToCart 에서 전역 지정이 불가능 하여 ProductDetail에 직접 지정
